Swapped cards. Deleted console.log
