+++ src/pages/Home/ProjectPortfolioSection/OldCard/OldCard.module.css
@@ -1,55 +1,46 @@
 .wrapper {
-  height: 360px;
+  display: flex;
+  max-width: 1200px;
+  max-height: 338px;
   width: 100%;
-  display: flex;
-  border-radius: 10px;
-  border: 1px solid rgba(252, 252, 252, 0.2);
-  transition: all 0.5s;
-}
-
-.wrapper:nth-child(2n) {
-  flex-direction: row-reverse;
-}
-
-.cardWrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  flex: 1;
-  flex-basis: 0;
-  padding: 30px;
-  border-radius: 10px;
-  transition: all 0.5s;
-}
-
-.headingWrapper {
-  display: flex;
+  height: 100%;
   justify-content: space-between;
 }
 
-.heading {
+.titleBlock {
+display: flex;
+flex-direction: column;
+justify-content: space-between;
+max-width: 250px;
+}
+
+.platformBlock {
+  display: flex;
+flex-direction: column;
+justify-content: space-between;
+align-items: flex-end;
+width: 210px;
+}
+
+.image {
+  position: absolute;
+  top: 5px;
+  left: calc(50% - 235.5px);
+  border-radius: 15px 15px 0px 0px;
+}
+
+.title {
   color: var(--white);
   font-size: var(--h4-headline-font);
   font-weight: var(--h4-headline-font-weight);
+  max-width: 210px;
 }
 
-.index {
-  color: var(--white);
-  font-size: var(--text-description-font);
-  font-weight: var(--text-description-font-weight);
-  opacity: 0.4;
-}
-
-.descriptionBlock {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.subtitle {
+.subTitle {
   color: var(--white);
   font-size: var(--h5-headline-font);
   font-weight: var(--h5-headline-font-weight);
+  margin-bottom: 10px;
 }
 
 .text {
@@ -59,69 +50,9 @@
   opacity: 0.6;
 }
 
-.label {
-  font-size: var(--text-secondary-font);
-  font-weight: var(--text-secondary-font-weight);
+.platform {
+  font-size: var(--text-main-font);
+  font-weight: var(--text-main-font-weight);
   color: var(--white);
   opacity: 0.6;
-  display: block;
 }
-
-.imgWrapper {
-  width: 40px;
-  height: 40px;
-  background: var(--white);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 50%;
-  fill: var(--black);
-}
-
-.imageWrapper {
-  flex: 1;
-  flex-basis: 0;
-  overflow: hidden;
-  padding: 30px;
-  position: relative;
-}
-
-.hoverImage {
-  position: absolute;
-  right: 0;
-  top: 0;
-}
-
-.wrapper:hover {
-  background: #252525;
-}
-
-.wrapper:hover .cardWrapper {
-  background: #202020;
-  border: initial;
-}
-
-.wrapper:hover .imgWrapper {
-  background: var(--green);
-}
-
-.wrapper:hover .imgWrapper :global(.icon__fill path) {
-  fill: var(--white);
-}
-/* .wrapper:hover .imgWrapper :global(.icon__fill path) {
-  fill: var(--white);
-} */
-
-/* .wrapper:hover .text,
-.wrapper:hover .label {
-  color: var(--black);
-} */
-
-/* .wrapper:hover a {
-  background: var(--orange);
-  color: var(--white);
-} */
-
-/* .wrapper:hover .image {
-  transform: scale(1.1);
-} */
