+++ src/pages/Home/ServiceSpectrumSection/BlueSectionWrapper.tsx
@@ -8,7 +8,6 @@ const BlueSectionWrapper = ({ children }: { children: React.ReactNode }) => {
 
   useEffect(() => {
     if (window) {
-      console.log('window');
       setWindowHeight(window.innerHeight);
     }
   }, []);
